pyproject.toml
[tool.poetry]
name = "mextractor"
version = "3.0.0"
description = "mextractor can extract media metadata to YAML and read them"
readme = "README.md"
repository = "https://github.com/caniko/media-metadata-extractor"
homepage = "https://pypi.org/project/mextractor/"
authors = ["Can H. Tartanoglu <canhtart@gmail.com>"]
license = "Apache-2.0"
keywords = ["pydantic", "metadata", "video", "image", "bigdata"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = ">=3.9, <3.12"

numpy = "*"

pydantic = "^1.9.2"
pydantic-numpy = "^2.2.1"

"ruamel.yaml" = "^0.17.21"
opencv-python = "^4.6.0"

ffmpeg-python = {version = "^0.2.0", optional = true}

click = "^8.1.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[tool.poetry.extras]
video-extract = ["ffmpeg-python"]

[tool.poetry.scripts]
mextractor = "mextractor.cli:mextractor_cli"

[tool.mypy]
python_version = "3.10"
plugins = "numpy.typing.mypy_plugin"

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
